requirements.txt
python-telegram-bot[job-queue]==20.7
anthropic==0.18.1
asyncio==3.4.3
aiofiles==23.2.1
python-dotenv==1.0.0
